refactor(gui): extract TimeSeriesGraphWidget to its own header (#2554) — Principle VIII.
The System Info dialog (#2554) needs the same time-series chart the network
diagnostics dialog draws, and a file-private class in
NetworkDiagnosticsDialog.cpp cannot be shared. Move it out unchanged.

Pure relocation: the class body is byte-identical, it has no Q_OBJECT and no
signals or slots, so no meta-object or moc consequences. kThrottleAmber moves
with it — the widget draws the throttle band with it, and the dialog keeps
using it for the badge, the state label and the fps-cap trace through the new
header. NetworkDiagnosticsDialog.h already forward-declared the class inside
namespace AetherSDR, so that declaration resolves as-is and the header needed
no change.
— authored by agent (Claude Code) on behalf of @skerker
